Traffick in ints sooner than in floats
Where numeric values are to be scaled by 10^-6 for printing, store the
scaling parameter as an integer like 6, not as a decimal like 0.000001,
because the integer might be wanted later. Here, 6 is preferred to -6
because the negation is implicit in the algorithm, but the |6| is data.

int(6) can always be converted reliably to 10^6, and 10.0^6 is reliably
integer-valued if nonstd::power() is used instead of a possibly flawed
std::pow(). OTOH, 10.0^-6 can have no exact binary representation, and
relying on even an optimal standard library to recover the exponent as
an exact integer will someday end in tears--immediately in the happiest
circumstance, or decades later, as in the present case.

Subtleties:
 - To determine the proper scaling for a set of columns, std::min()
   was applied to 0.001 and 0.000001, where std::max() must be applied
   to 3 and 6, in order for uniform scaling to make them all fit.
 - There is no need to assert that the multiplier is nonzero, because
   no value of the integral exponent will make it so.
 - The default scaling parameter, formerly a multiplier of 1.0, becomes
   an exponent of 0, of course.