Factor out max-power-of-10 parameter [383]
Made the threshold value for scaling into an argument. Its value is
hard-coded for now, and the effect of changing it has not been tested.

It would not make sense to treat the highest supported scaling power a
parameter: currency amounts beyond quintillions will not be needed for
the anticipated lifetime of lmi.