Add a failing unit test
This tricky statement:
    double widest = std::max
        (min_value * -10
        ,max_value
        );
doesn't do the right thing for
    1'999'999'999.0 min_value
     -999'999'999.0 max_value
because it violates the assumption that min_value <= max_value. That
precondition must hold, because std::minmax_element() produces those
values; but it should be asserted, because their connection to that
standard-library function is indirect.

Of course, no assertion can make the statement less tricky.