Reduce trickiness
The widest element (considering '-' as wide as any digit) may very well
be the greater of these:
  std::floor(min_value) * -10.0
  std::ceil (max_value)
for all combinations of maximum and minimum values, but it's not easy to
tell at a glance. It is not materially more expensive, but much clearer,
to adjust each value in the same manner (dependent on algebraic sign):
if it's 999.99, make it 1000; and if it's -999, make it 9999, because
both have four characters (which is still objectionably artificial).