Reorder statements for robustness
scale_unit_ is a simple function of scale_power_, and should therefore
be set whenever scale_power_ is set. Apparently the original version was
not actually wrong, because of the particular value to which scale_unit_
is initialized; but a reader who attempts to trace that logic rigorously
will endorse this improvement.