Improve efficiency
Ascertaining the extrema of all vectors, and then determining a scaling
factor suitable for the extrema extremorum, is more efficient than
determining distinct factors for each of several groups of vectors and
then selecting the highest such factor.

The change to 'bourn_cast_test.cpp' is voodoo that prevents a unit-test
failure; it will be discussed soon on the mailing list.