Reimplement scale_power()
Determine the formatted width of numeric column vectors (represented
by their minimum and maximum elements) by calling std::snprintf().

This is more robust than the std::log10() algorithm it replaces. In
this example:
  999,999,999.49 (log 8.999999999779) -->   999,999,999  9 digits
  999,999,999.50 (log 8.999999999783) --> 1,000,000,000 10 digits
it distinguishes small integers instead of comparing logarithms that
agree to within a relative error of 5e-13 (the example is germane even
though scale_power() rounds to whole integers). It also handles "-0"
and "inf" correctly.

std::snprintf() is much costlier than std::log10(), but the cost has
been mitigated by recent changes that first determine the minimum and
maximum of all scalable columns in a ledger. This cost is negligible in
the context where scale_power() is used, which formats and prints every
value for an illustration.