Avoid accidentally focusing wrong window in GUI tests
Use wxWindow::FindWindow(), which considers only the name of the window,
instead of FindWindowByName(), which falls back to the window label if
no window with the given name is found. This behaviour was unhelpful as
the function could find a wrong window (e.g. the static label preceding
the control instead of the control itself) if a label, instead of name,
was accidentally used as its argument.