Fix defect introduced 20170409T1141Z: impermissible constexpr
In commit 55d589680, various function templates were made constexpr,
conditional on "201402L < __cplusplus", which was not true for the
version of gcc then used in production, but has since become true
(the conditionals accordingly having been removed). However, the
floating-to-integral conversion could not actually be constexpr, e.g.
because it used std::ostringstream; now it does not. See:
  https://lists.nongnu.org/archive/html/lmi/2018-03/msg00085.html
et seqq.