Add pdf_writer_wx::save() to explicitly save PDF documents
Instead of always calling wxPdfDC::EndDoc() in pdf_writer_wx dtor, only
do it from the newly added save() method. This ensures that the PDF file
is not saved at all if an error occurs while generating it, instead of
creating an invalid (incomplete, and possibly empty) file.

Add various safety run-time checks to ensure that save() is called once
and once only and that no other methods of pdf_writer_wx are called
after it.

Update all code using pdf_writer_wx to call save(). For pdf_illustration
class this involved replacing "writer_" member variable with a "writer"
local one and, as consequence, passing the output file name parameter to
render_all() method instead of ctor -- which is more logical in any case
as only render_all() really creates the output and, in principle, could
be called multiple times with different file names on the same object.