Shorten diagnostics
The benefit of elaborate assertions is that they specifically indicate
what has gone wrong. The cost is that they make the code longer and
harder to read, and that they must be maintained if the variables and
functions they use ever change. It's a tradeoff, and a matter of taste,
but in this case simplicity seems preferable.