Consolidate and revise documentation
It seems generally best to comment a member function in one place only:
at its point of definition. This particular comment should be included
in documentation generated with a tool such as doxygen, so it's written
as a '///' block, with one blank line following, immediately preceding
the function definition.

The content of the original is preserved except for one deliberate
change: if the canonical
  ctor: N = std::uncaught_exceptions();
  dtor: if(N != std::uncaught_exceptions())
idiom had been used, and N had been 1 in both the ctor and the dtor,
then the dtor wouldn't omit a desirable though noncritical diagnostic:
instead, it would emit an unintended diagnostic.

It is arguable whether it would be simpler to follow that canonical
idiom rather than expending a larger number of characters in a comment
explaining why it wasn't followed.