Suppress warnings when building libxml2 et cetera
Building 'install_libxml2_libxslt.make' gives
  3736 lines of output before this change
  2061 lines of output after  this change
  1675 = 3736 - 2061
The warnings are assumed to be benign because lmi has built these
libraries the same way for years with no apparent problem. Suppressing
them is useful because cluttering build logs makes it harder to notice
diagnostics that really are significant. For example, in the log of a
recent 'install_msw.sh' run
  $wc log-20180411T0857Z
     6366   77499 1220816 log-20180411T0857Z
1675/6366 = 26% of the total lines are ignorable diagnostics from this
particular makefile alone.

Prefixed '-g -O2' to $CFLAGS overrides because that was the previous
behavior when $CFLAGS was not overridden:
  https://www.gnu.org/software/autoconf/manual/autoconf.html
| If output variable CFLAGS was not already set, set it to -g -O2 for
| the GNU C compiler

Eight '-Wsuggest-attribute=format' warnings remain despite the attempt
to inhibit them with '-Wno-suggest-attribute=format'.