Clone wx repository into cache if not yet cached
The newly-added code doesn't interact smoothly with the configurable
settings that follow it, because this commit deliberately changes the
essential logic of the script. Formerly, it cloned a local repository
from github by default, or from an alternative location (such as cache)
if available. Now, it ensures that a repository exists in the cache,
and clones a local repository from cache. Thus, git repositories are
treated the same as downloaded files:
 - if missing from cache, fetch from the internet into cache
 - always obtain only from cache for local use
IOW, the cache directory functions as a proxy.

In particular, $default_url is now unconditionally set, so there's no
point in overriding $wx_git_url.  We need to determine exactly what part
of $wx_git_url needs to be configurable.