No longer test PDF emission from command-line binary
The old PDF code will soon be expunged. The new PDF code depends on
wxPdfDoc, which is not available in non-wx-dependent builds, which in
turn will become unable to produce PDF output at all.

Incidentally, specified full paths to input files in the modified
commands, to make them work when pasted into a terminal.