Transplant implementation of a nested class
Forward-declared a nested class, and moved its implementation out of
line, to promote readability. This change is merely textual: the
nestedness of the inner class is preserved.

Ideally, the inner class's implementation would be moved into the same
TU as the outer class; that's not possible because the outer class has
a std::vector of the inner class's type. See:
  http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2015/n4371.html