Add a map of formatting masks to complement column titles
Copied title map to construct a new mask map with the same mask for all
columns (for the moment), so that this should be a pure refactoring.

Thought should be given to merging these maps with the format map. At
present, apparently the format map is used to format each datum in a
column, while the "999,..." mask (hard-coded elsewhere, and tabulated
in this commit but used only for supplemental reports) is used only to
determine a column's width.