During comprehensive non-GUI test, prompt to run GUI test
The GUI test should only be run manually, because by its nature it
requires full control of the console. But ideally it is run at the
same time as other automated tests that request all available CPUs.
However, it cannot be run before its binary has been built. This
change prompts an attentive operator when that binary has been built,
so that the GUI test can be run in parallel with all other tests if
desired.