Merge the last two 'skin.xrc' changes into 'skin_group_carveout.xrc'
This merge is in the spirit of the last commit to 'skin_coli_boli.xrc',
but with certain changes:
 - CountryCoiMultiplier is copied but commented out, because it really
   does belong here but might surprise end users; and
 - ProjectedSalary is omitted because, for unfortunate historical
   reasons, it's on an "Amounts" tab.
The combined effect of those changes, along with the oddly tall and
narrow geometry of this skin (attributable to the poor design of the
"Plan" tab), makes this "Participant" tab (and other tabs, e.g.,
"Amounts") look a little weird. The real solution to these issues
begins with cleaning up the "Plan" tab first.