Explicitly hardcode path to local wxxrc schema
Formerly, 'xrc.rnc' read the latest schema from wxwidgets.org each time
it was invoked. That contradicted the principle that lmi tools should
work correctly even without internet access, provided only that copies
of all online resources have been cached. It did ensure that the latest
version of the schema would always be used, but that is not necessarily
a benefit: the latest online version could be incompatible with lmi's
XRC files, either deliberately, or due to an upstream mistake.

Now, the schema in the current local working copy of wx is used instead.
It is guaranteed to be the version intended for use with the wx version
actually used to build lmi. Any enhancements introduced subsequently are
ignored, but so are any upstream errors or incompatibilities.

Alternatively, an xml catalog could have been used to cache the schema.
That would require installing a URI resolver that is not already present
in the deliberately minimal chroot or cygwin system used for production
builds. Installing a resolver would add a point of failure and enlarge
the attack surface: too much extra baggage for a single small file that
is already available locally and rarely changes.