Move most members from derived to base class
Historically, base class wx_table_generator was written to support group
premium quotes, and the later derived class illustration_table_generator
was "slightly specialized" to support illustration PDFs. In a textbook
design, the base class might provide only what's common between the two
use cases, each of which might then have its own derived class for
specializations:

             B (base)
             |
      +------+------+
      |             |
     DQ (quotes)    DI (illustrations)

For historical reasons, operations like output_highlighted_cell() that
are used only for group quotes were added to class wx_table_generator,
which was the only table-generator class at that time, before the idea
of reusing the code for illustrations was discussed. Therefore, that
class may seem to be B because it's used as a base of DI, but really it
performs the job of DQ--and deriving DI from DQ seems weird.

It would be possible to refactor class wx_table_generator into B and DQ,
but the differences are too small to justify an elaborate hierarchy, so
instead
 - this commit merges the differences DI-B into B, and
 - the next commit will eliminate DI (illustration_table_generator),
so that a single class will be used for all purposes.

This is only an initial top-level step. The historical context pervades
the design--e.g., the notional DQ centers all columns but one, and DI
overrides that to right-align all its columns, but there's solid B code
in wx_table_generator that is capable of handling alignment flexibly, if
only class wx_table_generator::column_info realized that.

This commit is mostly a copy-paste move, and
  git diff --ignore-all-space --color-moved=plain
makes that easier to see.