Generalize class wx_table_generator: row height and boolean flags
Chose settings in the ctor, depending on use case. Eliminated two public
member functions. In using_illustration_table::create_table_generator(),
constructed a wx_table_generator directly in the return statement.

Every member that had been initialized by a brace-or-equal-initializer
or a ctor-initializer is still initialized, but perhaps along one of the
ctor's two execution paths. Although this may be confirmed today, with
some effort, it's not robust under maintenance.