Generalize class wx_table_generator more robustly
Replaced switch-on-enum with tag dispatch. Wrote a separate ctor for
each use case. Initialized all scalar data members in ctor-initializer.

This approach preserves the readability of its predecessor in that all
initializations are grouped together for each use case. It improves
robustness by performing initialization in the ctor-initializer; that
is deemed to be worth adding about twenty extra lines.

The only data member not mentioned in a ctor-initializer is the vector,
which has its own default ctor.

It seems inoffensive enough to duplicate the first five lines of both
ctors; refactoring could of course turn those into a single line.