Fold a "ternary" conditional into a switch
The 'widest_text.empty()' condition has been replaced by the explicit
elastic-inelastic dichotomy. It is clearer to test the condition OAOO
and write one block for each set of conditionally-executed statements.

Wrote temporary assertions to ensure that the conditions are aligned
correctly. They'll be expunged without comment in the next commit.