Prefer 'n_'- prefix for "number of"
Generally changed '\<num_' prefixes to the shorter 'n_' that is already
widely used throughout lmi.

In 'input_sequence_entry.cpp', changed this:
    int const num_intervals = intervals.size();
which meant "number of intervals", but not this:
    int num_i = duration_num_field(i).GetValue();
which means "numerical duration value in the ith field".

Did not change 'rate_table*.cpp', which is a standalone utility suite
independent of the production system with its own self-consistent style.