Always derive numeric summary page from numbered_page
Simplify the page classes structure by not using a separate
attachment_page class nor ill_reg_numeric_summary_or_attachment_page
template, but just two plain ill_reg_numeric_summary_page and
ill_reg_numeric_summary_attachment classes.

It could be argued that this is not, strictly speaking, correct, as an
attachment page doesn't really have a page number, but in this code
"numbered" basically means "showing something like a page number in the
footer" and from this point view, this change makes more sense.

Moreover, in the upcoming changes, numbered_page will also be extended
to take care of splitting the output in several physical pages, if
necessary, and this functionality would also be useful to have for the
attachment pages as well.

Finally, this change makes the code simpler and more straightforward,
which is always welcome.