Resolve more '-Wconversion' issues
* bourn_cast.hpp: Intruded static_cast into the integral-to-floating
conversion, giving the rationale in a comment block. Incidentally
aligned range comparisons thus:
  if(x < y)
  if(    y < z)
to make it easier to see that 'y' is the same when it's a complicated
expression. Similarly, added a static_cast to the floating-to-integral
code, making it look more like 'narrow_cast' in TC++PL4 [11.5].

* bourn_cast_test.cpp: Suppressed compiler warnings with pragmata. The
static_cast workaround seemed less appealing here, in particular because
one test converts a value both explicitly (with static_cast) and
implicitly, then compares the converted values. (Perhaps that test can
never fail, but, this close to the edge of UB, it could conceivably find
a compiler defect.)