Resolve more '-Wconversion' issues
Used a pragma to ignore conversion warnings on a single statement.

As the comments indicate, this code was written for C++98, which assumed
only a C89 standard library. A thorough modernization would take a lot
of time and effort; that doesn't seem justifiable for code that has
worked perfectly since 2004 as far as we know.