Inhibit '-Wsign-conversion' for especially problematic files
It is difficult to use '-Wsign-conversion' with C++ code that uses the
STL because:

    std::vector<int> v(1);
    int i = 1;
    return v[i];
              ^
  error: conversion to 'std::vector<int>::size_type {aka unsigned int}'
  from 'int' may change the sign of the result [-Werror=sign-conversion]

See "Going Native 2013"--Stroustrup: "one of the sad things about the
standard library is that the indices are unsigned", and Sutter: "It's
unfortunately a mistake in the STL, in the standard library, that we use
unsigned indices".

However, it is useful occasionally to build with this flag, filtering
output with:
  sed -e '/size_type/d'
to identify remediable warnings.