Fix defect introduced 20111009T1501Z: wrong datatypes
wxDataViewListCtrl::GetItemCount() returns unsigned int, not long int.
(It's notionally a signed int because it's not a bitfield and mod 2^N
arithmetic is not wanted. It must be cast to long int because of the
wxVariant API, but that's a separate issue.) wxSpinCtrl takes integer
parameters, not long int. See:
  https://lists.nongnu.org/archive/html/lmi/2018-06/msg00004.html
Replaced static_cast with bourn_cast, which does the same thing if the
cast preserves value, but warns if it does not.