Transplant and refine built-in-type size assertions
Moved these static assertions to a file that's always included in every
binary, because lmi assumes them to be valid universally. To that end,
included 'main_common.hpp' in 'cpp_main.cpp', so that unit-test binaries
are also built with these assertions.

Dropped the 'short int' assertion: that type is unimportant in this
millennium. Improved the 'int' assertion: larger 'int' is okay.