Require specifying font sizes when using pdf_writer_wx
Make the previously optional "font sizes" argument mandatory.

This seems to make more sense because the 2 users of this class use it
with different font sizes sets and none of them is clearly better than
the other, so specifying it in one case and not in the other was
gratuitously asymmetric.

It also simplicities pdf_writer_wx ctor code and avoids duplicating the
font names in it.