Reimplement column_value_varies_across_cells()
The only sensible use case is to test whether any class default or any
cell has a different value than the case default for the given column.
To achieve that, the function had to be called twice--once for classes,
and once for cells--but it was called incorrectly in one of the two
cases where it was used. The documentation reflected the sensible
intention, and was therefore wrong.

Reimplemented the function. Incidentally used a more descriptive name
for an entity specified by a for-range-declaration's declarator (it
would be nice to have a shorter technical term for such a thing), and
replaced !(x==y) by x!=y. Left the old implementation in place for the
moment only to assert that new and old are equivalent.