Fix defect introduced 20180228T1615Z: bad output filename
When this unreleased experimental code was introduced, the name of a
pasted-out census file (assuming the spreadsheet extension is '.tsv')
was formed thus:
  xyz.cns      --> xyz_pasted_out.tsv
which wasn't egregious, except...
  client.1.cns --> client.tsv  ['.1' infix lost]
  client.2.cns --> client.tsv  [different input name, same output name]
Moreover, infixes added to filenames elsewhere are dot-delimited:
  xyz.cns --> xyz.000000001.pdf
and similarly for ".monthly_trace", ".quote", and ".roster". To follow
the same pattern, the infix should begin with '.', not '_'.

Fixed these defects in preparation for initial production release. The
examples above now work appropriately:
  xyz.cns      --> xyz.pasted_out.tsv
  client.1.cns --> client.1.pasted_out.tsv
  client.2.cns --> client.2.pasted_out.tsv