Use more recent wxPdfDocument fixing the text origin bug
Upgrade to the latest master version (identified as v0.9.6-28-gf66b428
by "git describe") of wxPdfDocument which includes the fix to the text
origin bug in wxPdfDC which misinterpreted the vertical position passed
to DrawText() method as being the text baseline instead of its
upper-left corner, as it should have been taken according to wxDC API
(see https://github.com/utelle/wxpdfdoc/pull/42 for more details).

This requires slightly updating the makefile building the library, as
its build system has changed in the meanwhile; and also updating all the
code which used to work around the now fixed bug and doesn't have to do
it any longer: we can now remove the different workarounds that were
used previously for positioning text correctly, notably the hack with
using an extra row in an HTML table in group quotes generation code.