Merge first two notes pages of regular individual illustrations
The separation of the first part of explanatory notes in 2 different
pages was a work around for lack of pagination in the illustration
generation code and doesn't need to be done any longer now that
pagination is supported, so merge the contents of reg_d_indiv_notes2.mst
template into reg_d_indiv_notes1.mst and remove the latter.

However the file with this name still exists, as it is the new name of
reg_d_indiv_notes3.mst, which was renamed to keep the numbers
consecutive, and kept separate for compatibility with the XSL-FO-based
illustrations where the contents of this template was explicitly kept on
a separate page.