Use standard <img> tag instead of <scaled_image> in templates
We can replace the standard <img> wxHTML tag handler with our own one
for the same tag instead of defining a custom tag and this looks
preferable, as, even if we still use a non-standard attribute, the <img>
tag is still understood by browsers and other programs working with HTML
(e.g. syntax-highlighting editors), unlike <scaled_image>.

There are no changes to the generated illustrations.