Exercise CC as well as CXX
Motivation: Additional gcc warnings will soon be enabled. It seems
desirable to maintain distinct lists for gcc, g++, and both. In the
future, lmi might use C files, as it did in the past, but no longer
does until this commit; adding one now ensures that the warnings are
all tested. And it may be more efficient to compile glibc code as C.

* nychthemeral_test.sh: Filter out CC output. Incidentally immunize
this and other filters against superfluous whitespace.
* round_glibc.cpp: Renamed to round_glibc.c .