Experimentally count copied cells on statusbar
Added code to help evaluate the idea that displaying an iteration count
on the statusbar is better than showing a busy cursor. If the entire
operation is very fast, then flashing a busy cursor is just a nuisance.
Showing a count on the statusbar serves the same purpose and is less
obtrusive. However, if each iteration is very fast, then the overhead of
displaying progress on the statusbar may be significant: in this case,
it makes a half-second operation take approximately twice as long.