Add a new element for input version 9
Added <AdditionalReports>, which is presently unused. This may someday
replace the unused <SupplementalIllustrationType>, which permits only an
enumerated set of extra reports. Being of a general string type, the new
element allows an extensible set of reports, which might someday be
designed and stored by end users. "Additional" in the name distinguishes
new reports from the "Supplemental" report already supported.