Strictly forbid inforce illustrations in month of issue
Formerly, this was forbidden only for input files from external sources.
The (flawed) reasoning was that entering an infelicitous set of dates
made it impossible to edit the file and correct the error. However, only
Input::set_inforce_durations_from_dates() prevents such manual recovery;
the code changed in this commit never did.