Force input inforce-as-of date to be an exact monthiversary date
This experimental change warns when the input inforce-as-of date is not
an exact monthiversary date, and then forces it to be so. This is the
same as the behavior prior to 2015-11, except that now a warning is
displayed in all cases, and no input file is summarily rejected.

Changes in 2015-11 gave an error message essentially identical to this
commit's warning, and terminated the current task. This was found to be
too severe, because it prevented editing an offending cell to correct
the problem. For that reason, the date check was bypassed (and any
problem was not even diagnosed) except for input files from external
sources.

Now, several years later, the problem is not seen in external files: the
diagnostic code provides peace of mind, but its message is not observed
in practice. However, input entered in lmi has continued to be treated
as it was before 2015-11.

The new behavior might not be ideal, and may later be refined. The
immediate goal is to remove the dependency on Input::InforceDataSource,
which needs to be expunged.