Fix warning about comparing double with bool in group quotes code
LedgerInvariant::GroupIndivSelection field has type double and was first
implicitly converted to group_quote_pdf_gen_wx::individual_selection of
type bool, which was already not very clear but didn't trigger any
warnings, and was also compared with bool, which did result in a warning
from MSVS about unsafe comparison.

Fix the warning by converting the double value to boolean once, and
explicitly, and then only compare booleans with each other.

Alternatively, the type of individual_selection_ could be changed to
"double", but it seems better to keep it as boolean, as this is what it
really is.