Call Reconcile() when reading all external input files
Previously, this was done only for '.cns' input; now, it's done for
'.ill' input as well. (Others like '.gpt' never come from external
sources.) This change may seem pointless per se, because loading an
'.ill' file immediately opens the tabbed dialog, thereby triggering
Reconcile() (and external files are not customarily used with CLI
builds that have no dialog to open). However, it isn't a bad idea to
handle both types of (external) files in the same way now, and soon
it will become a very good idea because an additional validation step
is about to be added.