Preserve configurable settings after GUI tests execution
Make a backup copy of the configurable_settings.xml file changed by the
GUI unit test and restore it after the test end.

The copy is made on disk, which makes the code much more complicated
than if we simply kept it in memory, however it is safer because the
temporary disk file remains even if the test crashes or otherwise
misbehaves (or even, more prosaically, simply killed while debugging).