Rename class wx_table_generator::column_info --> table_column_info
Motivation: To unit-test compute_column_widths(). Its diagnostic
messages frequently appear, indicating either a regression such as:
  https://lists.nongnu.org/archive/html/lmi/2018-07/msg00020.html
or design issues that must be grappled with, e.g.:
  https://lists.nongnu.org/archive/html/lmi/2018-04/msg00100.html
and it is hoped that a rewrite can address the lingering questions
documented inline and lead to exploration of alternatives. This
planned work is much easier and safer with unit tests.

compute_column_widths() is already independent of any wx code, but it
uses the class that is here renamed and moved outside of a wx-dependent
enclosing class; soon these things will be moved into new source files.