Change schema's "ListBillDate" default value
Made this element's default value the same in the schema as in lmi
itself, i.e., JDN 2440588 = 1970-01-01: a date before which it is
assumed that no list bill need ever be generated. A value that is
implausible on the face of it is actually advantageous: it reminds
users who really want a list bill to enter a billing date explicitly,
while for others it is more readily disregarded. Prescribing this same
value for external systems avoids suggesting to users that the value
should vary across cells, and makes it harder for vendors to get wrong.