Add a new element for input version 9
Added <IsInforce>, which is presently unused.

The problem that this may someday solve is that at present a cell is
considered to be inforce iff
    EffectiveDate != InforceAsOfDate
but, for new business, it is all too easy to make a cell "inforce" by
accident, e.g., by decrementing 'EffectiveDate' while leaving
'InforceAsOfDate' unchanged.

Release candidate 20180719T1706Z can be put into production without
this change, without causing any harm. Files saved with that version
will lack <IsInforce>, which will therefore be given a default value
that is not used. The schema will consider these files defective, but
files created by lmi are not subject to schema validation, so that
doesn't matter.