Remove monthly trace file created by output validation unit test
This file was created as a side effect of running the cell in the test
case for copying illustration data and was never removed.

Add an output_file_existence_checker testing that the file was indeed
created (which is not really required, but doesn't seem like it could
cause any harm) and, more importantly, ensuring that it is removed at
the test end.