Clean up extra files created during MEC and GPT UI tests
Creating a new MEC or GPT document results in creation of 2 or 1 extra
file in the current directory, which were not checked for nor cleaned up
by the testing code.

Do add both the checks (that might be unnecessary) and cleanup for them
now.

To simplify the new code, make output_file_existence_checker class
movable and add path() accessor to it.