Prescribe an appropriate value for <IsInforce>
Formerly, the schema said to default this element's value to 'Yes',
which seemed appropriate because in practice files are provided by
external systems only for inforce. However, that conflicted with the
advice, in 'multiple_cell_document.rnc', to copy case and class default
cells from this file:
  http://git.savannah.gnu.org/cgit/lmi.git/plain/sample.cns
which are not inforce.