Prevent page breaks inside paragraphs in a more robust way
Just marking the current container as unbreakable in <P> HTML tag
handler worked as long as the paragraph contained only text and inline,
span-like tags such as <B> or <I> -- which was indeed the case in the
current HTML templates, ensuring that pagination worked correctly.

However a paragraph could still be broken if it contained any
non-span-like element, such as <OL> or <UL> or even a simple <BR>.
Prevent this from happening by keeping all such elements inside a nested
wxHtmlContainerCell, which is not closed until the end of the paragraph,
and thus ensuring that the parent container, corresponding to the
paragraph itself, remains unbreakable in any case.