Generally prefer 'int' to other integer types
Replaced 'std::size_t' with 'int' in the private member functions of
class wx_table_generator. Eventually, it may be replaced by a class type
or a scoped enumeration in the public interface, with the noteworthy
advantage that (unlike std::size_t) such types are not implicitly
intraconvertible with 'int'.