Refactor for clarity
Passed actual total margin instead of half the total. Now the margin
argument means the space between columns on an illustration, where all
columns are inelastic and right-aligned anyway. This code was originally
written with group quotes in mind, where all columns but one are
centered, and margins were therefore bilateral--one em on each side:
  emspace data emspace
However, for illustrations, there is no marginal space on the right:
  case oe_right:
    x_text += ci.col_width() - dc().GetTextExtent(s).x;
    dc_.DrawText(s, x_text, y_text);
thus:
  emspace emspace data
and it is more natural to think of the two-em space between columns as
the "margin".