Assert an apparent requirement
Experimentally, in never-to-be-committed local changes, changed spacing
between columns to 7em, and observed some "super" headers (those that
span multiple columns) superimposed one over another. Because this has
not been observed in production, concluded that they must all fit within
disjoint bounding boxes. Asserted this, because it is important to know
if this invariant is broken by any future change.