Use braces instead of parentheses in the initializer lists
No real changes, simple use C++11 "uniform initialization" syntax for
the members and base classes initialization, for consistency with the
field initializers and because this kind of initialization is safer as
it avoids narrowing implicit conversions.

Take care to _not_ change parentheses to braces in the initialization of
std::vector or std::deque fields as, for them, making this change can
change the kind of ctor being used from the one taking the initial
container size and the value of its elements to the one taking
std::initializer_list.