Fix the problem analyzed in the penultimate commit
The code being worked on contained a block comment that suggested a
design regret, so the time is ripe to "refactor mercilessly". This
commit includes a proposed simplification for one function, which is
for the moment asserted to do the same thing as the original; after
testing, the original will be removed, in the next commit.