Pass the illustration and interpolator to the page ctor
No real changes, just refactor the page classes to take the illustration
object directly in the ctor, instead of using a separate illustration()
method called immediately after the page is created, and also passing
the interpolator to the ctor as well.

This makes things simpler as the interpolator is now consistently
available in all page code and doesn't need to be passed around all the
time and doesn't really compromise locality as the interpolator is used
via a const reference which previously was made accessible to any code
that needed it anyhow.