Avoid parsing the same HTML more than once in standard pages
Add pdf_writer_wx::parse_html() which can be used to parse HTML into the
internal representation and change paginate_html() and output_html()
used for multi-page output to take this internal representation instead
of the text itself.

This avoids parsing the same HTML twice, which is relatively
time-consuming: for the total illustration generation time of ~1000ms,
parsing HTML took ~200ms before this change and ~120ms after it,
resulting in 8% speed up.

Moreover, this will make it possible to modify the HTML text before
rendering it in the upcoming commit, which would have been difficult and
awkward to do if we had to continue passing HTML text to output_html().