Implement support for headers for normal PDF illustration pages
The <header> tag can now be used in HTML templates to define the
contents of a header which will be repeated on any subsequent physical
pages if the logical page contents doesn't fit into a single one.

Show this tag in action in the "Explanatory Notes" page of reg_d
individual illustration: with sample2ipp its second page correctly
repeats the logo and the title now, while it didn't have them before.